Extract Store recovery methods to impl module.
Follows 1372d0493afbf3926a390b023ad11114c28efc99 this time further
shrinking the store to just over 2K lines by moving recovery-related
methods into their own impl.
